Initial Ideas
Basic one button window with Tkinter to test the concept. More buttons next.
I May look into styling, but since this is a functional project to be used behind the scenes, that should only be necessary to make it look more user-friendly when there are more sounds.

Ideal final product is something that will contain the files for all the shows, and can be navigated per show.